Move HighlightsState and getPreloadPages to model.ts
Relocate HighlightsState type and getPreloadPages function from
PageViewer.tsx into the shared model.ts module. This allows them
to be imported by the new usePageFind hook and other consumers
without creating a circular dependency on PageViewer.

Update PageViewer.tsx and Controls.tsx imports accordingly.
